[docs] 프로토콜 파일 추가
