+++ DrawingApp/DrawingApp/RandomViewFactory.swift
@@ -1,0 +1,4 @@
+protocol RandomViewFactory {
+    init(deviceSafeArea: Point, deviceScreen: Size)
+    func makeView() -> RandomFigureView
+}
