+++ DrawingApp/DrawingApp/RectangleViewFactory.swift
@@ -1,10 +1,5 @@
 import Foundation
 
-protocol RandomViewFactory {
-    init(deviceSafeArea: Point, deviceScreen: Size)
-    func makeView() -> RandomFigureView
-}
-
 class RectangleViewFactory: RandomViewFactory {
     // init으로 받도록 변경한 이유는 어차피 하나의 장치 내에서 사용하므로
     // safeArea는 한번만 초기화 하면 이후에는 계속 재사용이 더 효율적이라 판단함
